Update Solex_recon.py
Proposed code to output a solar disc of a particular size (e.g. 900 pixel radius, like GONG) *AND* put it in a canvas of a given width (e.g. 2400 pixels).  Use case is to enable easier stitching together of sweeps to create a larger image in cases where full disc image is not available to the system.